+++ tests/deno_test_runAll.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-deno run --import-map=import_map.json --allow-read deno_test_runAll.js
+deno run --import-map=import_map.json --allow-read --allow-net deno_test_runAll.js