+++ tests/test_pcg32_implementations.js
@@ -2,7 +2,6 @@
 import {
   PluggablePRNG,
   RandomGenerator_Pcg32,
-  RandomGenerator_Pcg32_alt,
   RandomGenerator_Pcg32_alt_bigint,
   SeedInitializer_Uint64
 } from '../source/pluggablePrng.js'
@@ -14,14 +13,9 @@ const seedGen = new PluggablePRNG({
 
 for (let i=0; i<50_000; i++) {
   const seed = await seedGen.randomUint32()
-  const pcg32_longfn = new PluggablePRNG({
-    seed,
-    RandomGenerator: RandomGenerator_Pcg32,
-    SeedInitializer: SeedInitializer_Uint64
-  })
   const pcg32_uint64 = new PluggablePRNG({
     seed,
-    RandomGenerator: RandomGenerator_Pcg32_alt,
+    RandomGenerator: RandomGenerator_Pcg32,
     SeedInitializer: SeedInitializer_Uint64
   })
   const pcg32_bigint = new PluggablePRNG({
@@ -31,7 +25,5 @@ for (let i=0; i<50_000; i++) {
   })
   const correct = await pcg32_bigint.randomUint32()
   const  uint64 = await pcg32_uint64.randomUint32()
-  const  longfn = await pcg32_longfn.randomUint32()
   if (uint64 != correct) throw Error(`${correct} != ${uint64} (uint64)`)
-  if (longfn != correct) throw Error(`${correct} != ${longfn} (longfn)`)
 }